requirements.txt
opencv-python~=4.10.0.84
numpy~=2.1.3
matplotlib~=3.9.2
pillow~=11.0.0